refactor: fold sim-helpers into simulation.ts
The two-tier split was clean: simulation-api.ts wraps the raw HTTP
surface; simulation.ts holds the high-level encapsulation
(SimulationBuilder + session-bound helpers). sim-helpers.ts was a
third tier without a real architectural justification — callContract /
getStxBalance / getFtBalance / getNonce / readDataVar are just more
session-bound wrappers in the same vein as SimulationBuilder.

No public-API change: index.ts re-exports the same names; the helpers
are still importable as `import { callContract, ... } from 'stxer'`.
